Update buffered lexer interface, change its usage + refactoring
